arch: x86: Fix BIOS reboot for CPUs with LA57 support
On platform with X86_FEATURE_LA57 enabled, the CR4.LA57 is set and
5-level paging is used. Upon reboot, if the chosen reboot_type is
BOOT_BIOS, the CR4.PCIDE bit is cleared and the machine doesn't go
any further.

The reason for this failure is the sequence following the clearing of
CR4.PCIDE bit. The CR3 points to the trampoline_pgd entry, which does
not expect 5-level paging, therefore the failure.

That's why both CR4.PCIDE and CR4.LA57 must be cleared atomically, so
that we don't run into such issue.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>