virtio-mem: virtio_mem_init: Access bb_size just in BBM mode
/* In BBM, we also want at least two big blocks. */
vm->offline_threshold = max_t(uint64_t, 2 * vm->bbm.bb_size,
			      vm->offline_threshold);
This line does not modify vm->offline_threshold depending on the data in
vm->sbm that shares this address is 0 in SBM mode.
I think it might be difficult to make sure when we change this in the
future.

This commit adds an if to make sure that this line just be executed in
BBM mode.

Signed-off-by: Hui Zhu <teawaterz@linux.alibaba.com>